+++ packages/solUtils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @moralisweb3/sol-utils
 
+## 2.17.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/common-sol-utils@2.17.0
+
 ## 2.16.1
 
 ### Patch Changes